+++ include/casm/monte/checks/io/json/CompletionCheck_json_io.hh
@@ -18,8 +18,13 @@ template <typename StatisticsType>
 struct CompletionCheckParams;
 
 /// \brief Construct CompletionCheckParams<BasicStatistics> from JSON
-inline void parse(InputParser<CompletionCheckParams<BasicStatistics>> &parser,
-                  StateSamplingFunctionMap const &sampling_functions);
+void parse(InputParser<CompletionCheckParams<BasicStatistics>> &parser,
+           StateSamplingFunctionMap const &sampling_functions);
+
+/// \brief Convert CompletionCheckParams<BasicStatistics> to JSON
+jsonParser &to_json(
+    CompletionCheckParams<BasicStatistics> const &completion_check_params,
+    jsonParser &json);
 
 /// \brief CompletionCheckResults to JSON
 template <typename StatisticsType>
@@ -116,17 +121,15 @@ void _parse_components(
       (parser.self.find_at(option / "component_index") != parser.self.end());
   bool has_name =
       (parser.self.find_at(option / "component_name") != parser.self.end());
-  if (has_index && has_name) {
-    parser.insert_error(option,
-                        "Error: cannot specify both \"component_index\" and "
-                        "\"component_name\"");
-  } else if (has_index) {
+  if (has_index) {
     _parse_component_index(parser, option, function, precision,
                            requested_precision);
-  } else if (has_name) {
+  }
+  if (has_name) {
     _parse_component_name(parser, option, function, precision,
                           requested_precision);
-  } else {
+  }
+  if (!has_index && !has_name) {
     // else, converge all components
     for (Index index = 0; index < function.component_names.size(); ++index) {
       requested_precision.emplace(
@@ -363,6 +366,54 @@ inline void parse(InputParser<CompletionCheckParams<BasicStatistics>> &parser,
     parser.value = std::make_unique<CompletionCheckParams<BasicStatistics>>(
         completion_check_params);
   }
+}
+
+/// \brief Convert CompletionCheckParams<BasicStatistics> to JSON
+inline jsonParser &to_json(
+    CompletionCheckParams<BasicStatistics> const &completion_check_params,
+    jsonParser &json) {
+  // TODO: write out calc_statistics_f parameters
+  //  to_json["calc_statistics_f_confidence"] = 0.95;
+  //  to_json["calc_statistics_f_weighted_observations_method"] = 1;
+  //  to_json["calc_statistics_f_n_resamples"] = 10000;
+
+  json["cutoff"] = completion_check_params.cutoff_params;
+
+  json["convergence"] = jsonParser::array();
+  for (auto const &pair : completion_check_params.requested_precision) {
+    auto const &key = pair.first;
+    auto const &req_prec = pair.second;
+
+    jsonParser tmp;
+    tmp["quantity"] = key.sampler_name;
+
+    std::vector<int> component_index;
+    component_index.push_back(key.component_index);
+    tmp["component_index"] = component_index;
+
+    std::vector<std::string> component_name;
+    component_name.push_back(key.component_name);
+    tmp["component_name"] = component_name;
+
+    to_json(req_prec, tmp);
+
+    json["convergence"].push_back(tmp);
+  }
+
+  // "spacing"
+  if (completion_check_params.log_spacing == false) {
+    json["spacing"] = "linear";
+    json["begin"] = completion_check_params.check_begin;
+    json["period"] = completion_check_params.check_period;
+  } else {
+    json["spacing"] = "log";
+    json["begin"] = completion_check_params.check_begin;
+    json["base"] = completion_check_params.check_base;
+    json["shift"] = completion_check_params.check_shift;
+    json["period_max"] = completion_check_params.check_period_max;
+  }
+
+  return json;
 }
 
 /// \brief CompletionCheckResults to JSON